pq infernos n ta pegando?

+++ Header Files/ContaPoupanca.h
@@ -4,10 +4,10 @@
 #define CONTAPOUPANCA_H
 class ContaPoupanca:public Conta{
     private:
-        int jurosExtra;
+        float jurosExtra;
     public:
       //RETIRE AS CHAVES QUANDO FOR DEFINIR O CONSTRUTOR
-      ContaPoupanca(const Usuario & = {"0",0,0,{"0","0","0","0"}}, const int &  = 0, const float & = 0);
+      ContaPoupanca(const Usuario & = {"0",0,0,{"0","0","0","0"}}, const float &  = 0, const float & = 0);
       void setJurosExtra(const int &);
       void CalcularJurosExtra();
 };
